+++ packages/test-list-tile/src/tests-generic.js
@@ -46,6 +46,22 @@ export default ({ ListTile, Icon, renderer: h, keys }) => {
       }
     },
     {
+      name: "Option: subContent",
+      component: ListTile,
+      attrs: {
+        title: "Ancillary Justice",
+        subContent: h("div",
+          {
+            style: {
+              background: "#c5e1a5",
+              marginTop: "12px"
+            }
+          },
+          "The body lay naked and facedown, a deathly gray, spatters of blood staining the snow around it. It was minus fifteen degrees Celsius and a storm had passed just hours before."
+        )
+      }
+    },
+    {
       name: "Option: events",
       interactive: true,
       component: ListTile,